webapp/frontend/requirements.txt
streamlit
streamlit_drawable_canvas
numpy
pillow
matplotlib
opencv-python-headless